+++ src/LightBDD.Core/Execution/IRunnableScenario.cs
@@ -1,7 +1,11 @@
+using System;
 using System.Threading.Tasks;
 
 namespace LightBDD.Core.Execution
 {
+    /// <summary>
+    /// Interface representing runnable scenario.
+    /// </summary>
     public interface IRunnableScenario
     {
         /// <summary>
@@ -9,6 +13,7 @@ namespace LightBDD.Core.Execution
         /// Any exceptions thrown in scenario steps will be wrapped in <see cref="ScenarioExecutionException"/>. Code calling this method can rethrow the original exception by calling <code>ex.GetOriginal().Throw()</code>
         /// </summary>
         /// <exception cref="ScenarioExecutionException">Thrown when any exception occurs during execution.</exception>
+        /// <exception cref="InvalidOperationException">Thrown when scenario was already executed.</exception>
         Task ExecuteAsync();
     }
 }
