+++ src/LightBDD.Framework/Scenarios/IIntegratedScenarioBuilder.cs
@@ -1,0 +1,16 @@
+﻿using LightBDD.Core.Extensibility;
+
+namespace LightBDD.Framework.Scenarios
+{
+    /// <summary>
+    /// Interface offering access to core version of scenario builder.
+    /// </summary>
+    /// <typeparam name="TContext"></typeparam>
+    public interface IIntegratedScenarioBuilder<TContext> : IScenarioRunner<TContext>
+    {
+        /// <summary>
+        /// Returns core version of scenario builder.
+        /// </summary>
+        ICoreScenarioBuilder Core { get; }
+    }
+}
