+++ src/LightBDD.Framework/Scenarios/IScenarioBuilder.cs
@@ -1,7 +1,4 @@
-﻿using LightBDD.Core.Extensibility;
-using System.Threading.Tasks;
-
-namespace LightBDD.Framework.Scenarios
+﻿namespace LightBDD.Framework.Scenarios
 {
     /// <summary>
     /// Interface allowing to compose scenario in fluent way.
@@ -9,19 +6,10 @@ namespace LightBDD.Framework.Scenarios
     /// <typeparam name="TContext"></typeparam>
     public interface IScenarioBuilder<TContext>
     {
-        IIntegratedScenarioBuilder<TContext> Integrate();
-    }
-
-    public interface IScenarioRunner<TContext> : IScenarioBuilder<TContext>
-    {
         /// <summary>
-        /// Runs test scenario by executing all specified steps.
+        /// This method should not be used by LightBDD users, but code extending LightBDD capabilities.<br/>
+        /// Returns core version of the builder, allowing to configure scenario.<br/>
         /// </summary>
-        Task RunAsync();
-    }
-
-    public interface IIntegratedScenarioBuilder<TContext> : IScenarioRunner<TContext>
-    {
-        ICoreScenarioBuilder Core { get; }
+        IIntegratedScenarioBuilder<TContext> Integrate();
     }
 }
